fix vulkan layer loading for macos
use the same VK_LAYER_EXPORT as in linux and android as the vulkan loader couldn't find the symbols in the layer, add VK_KHR_PORTABILITY_ENUMERATION extension to the supported extensions as the vulkan driver on macOS is not fully  conformant Vulkan implementation. Add the flags parameter in VkInitParams as VK_INSTANCE_CREATE_ENUMERATE_PORTABILITY_BIT_KHR should be used in case if VK_KHR_PORTABILITY_ENUMERATION extension is being enabled.